deployments/requirements.txt
supabase==2.0.0
python-dotenv==1.0.0
langchain-google-genai==1.0.0
langgraph==0.2.16
langchain==0.2.0
requests==2.31.0
typing-extensions==4.5.0
pydantic==1.10.0
